fix: changed calc function location and changed use clinical details usage
