+++ client/src/components/App/Content/InvestigationForm/TabManagement/ClinicalDetails/useClinicalDetails.ts
@@ -2,17 +2,18 @@ import axios from 'axios';
 import React, {useState} from 'react';
 import { useSelector } from 'react-redux';
 
+import theme from 'styles/theme';
 import logger from 'logger/logger';
 import {Severity} from 'models/Logger';
 import StoreStateType from 'redux/storeStateType';
 import IsolationSource from 'models/IsolationSource';
-import { useDateUtils} from 'Utils/DateUtils/useDateUtils';
+import { useDateUtils } from 'Utils/DateUtils/useDateUtils';
 import useCustomSwal from 'commons/CustomSwal/useCustomSwal';
 import { setFormState } from 'redux/Form/formActionCreators';
 import FlattenedDBAddress, {initDBAddress} from 'models/DBAddress';
-import useSymptomsUtils from 'Utils/ClinicalDetails/useSymptomsUtils';
 import { setIsLoading } from 'redux/IsLoading/isLoadingActionCreators';
 import ClinicalDetailsData from 'models/Contexts/ClinicalDetailsContextData';
+import { getDatesToInvestigate } from 'Utils/ClinicalDetails/useSymptomsUtils';
 import { setSymptomsExistenceInfo } from 'redux/Investigation/investigationActionCreators';
 
 import ClinicalDetailsSchema from './ClinicalDetailsSchema';
@@ -50,15 +51,15 @@ const useClinicalDetails = (parameters: useClinicalDetailsIncome): useClinicalDe
     const { id, setSymptoms, setBackgroundDiseases, didSymptomsDateChangeOccur } = parameters;
 
     const { alertError } = useCustomSwal();
-    const { getDatesToInvestigate, convertDate } = useDateUtils();
-    const { alertSymptomsDatesChange } = useSymptomsUtils();
+    const { convertDate } = useDateUtils();
+    const { alertWarning } = useCustomSwal();
 
     const epidemiologyNumber = useSelector<StoreStateType, number>(state => state.investigation.epidemiologyNumber);
     const investigatedPatientId = useSelector<StoreStateType, number>(state => state.investigation.investigatedPatient.investigatedPatientId);
     const tabsValidations  = useSelector<StoreStateType, (boolean | null)[]>(store => store.formsValidations[epidemiologyNumber]);
     const address = useSelector<StoreStateType, FlattenedDBAddress>(state => state.address);
+    const validationDate = useSelector<StoreStateType, Date>(state => state.investigation.validationDate);
     
-    const [coronaTestDate, setCoronaTestDate] = useState<Date | null>(null);
     const [isolationAddressId, setIsolationAddressId] = useState<number | null>(null);
     const [isolationSources, setIsolationSources] = React.useState<IsolationSource[]>([]);
     const [didDeletingContactEventsSucceed, setDidDeletingContactEventsSucceed] = React.useState<boolean>(true);
@@ -67,7 +68,6 @@ const useClinicalDetails = (parameters: useClinicalDetailsIncome): useClinicalDe
         getSymptoms();
         getBackgroundDiseases();
         getIsolationSources();
-        getCoronaTestDate();
     }, []);
 
     const getSymptoms = () => {
@@ -83,21 +83,6 @@ const useClinicalDetails = (parameters: useClinicalDetailsIncome): useClinicalDe
         }
         );
     };
-
-    const getCoronaTestDate = () => {
-        const fetchCoronaTestDateLogger = logger.setup('Fetching corona test date');
-        fetchCoronaTestDateLogger.info('launching corona test date post request', Severity.LOW);
-        axios.get(`/clinicalDetails/coronaTestDate`).then((res: any) => {
-            if (Boolean(res?.data)) {
-                fetchCoronaTestDateLogger.info('got results back from the server', Severity.LOW);
-                setCoronaTestDate(convertDate(res.data.coronaTestDate));
-            } else {
-                fetchCoronaTestDateLogger.warn('got status 200 but wrong data', Severity.HIGH);
-            }
-        }).catch(err => {
-            fetchCoronaTestDateLogger.warn(`Got 500 from coronaTestDate request: ${err}`, Severity.HIGH);
-        });
-    }
 
     const getBackgroundDiseases = () => {
         const getBackgroundDiseasesLogger = logger.setup('Fetching Background Diseases');
@@ -199,31 +184,29 @@ const useClinicalDetails = (parameters: useClinicalDetailsIncome): useClinicalDe
 
     const deleteIrrelevantContactEvents = (symptomsStartDate: Date | null, doesHaveSymptoms: boolean) => {
         const deleteIrrelevantEventsLogger = logger.setup('Deleting irrelevant contact events');
-        if(Boolean(coronaTestDate)) {
-            const allDatesToInvestigate = getDatesToInvestigate(doesHaveSymptoms, symptomsStartDate, coronaTestDate);
-            if(allDatesToInvestigate.length > 0) {
-                deleteIrrelevantEventsLogger.info('Sending to server date to delete contact events by', Severity.LOW);
-                setIsLoading(true);
-                axios.delete('/intersections/deleteContactEventsByDate', {params: {earliestDateToInvestigate: allDatesToInvestigate[0]}}).then((result) => {
-                    if(result.data?.data?.deleteContactEventsBeforeDate) {
-                        deleteIrrelevantEventsLogger.info('Deleting contact events finished with success', Severity.LOW);
-                    } else {
-                        deleteIrrelevantEventsLogger.error(`Deleting contact events finished with errors: ${result.data}`, Severity.LOW);
-                        alertError(deletingContactEventsErrorMsg);
-                        setDidDeletingContactEventsSucceed(false);
-                    }
-                }).catch(err => {
-                    deleteIrrelevantEventsLogger.error(`Failed to delete irrelevant contact events: ${err}`, Severity.LOW);
+        const allDatesToInvestigate = getDatesToInvestigate(doesHaveSymptoms, symptomsStartDate, validationDate);
+        if(allDatesToInvestigate.length > 0) {
+            deleteIrrelevantEventsLogger.info('Sending to server date to delete contact events by', Severity.LOW);
+            setIsLoading(true);
+            axios.delete('/intersections/deleteContactEventsByDate', {params: {earliestDateToInvestigate: allDatesToInvestigate[0]}}).then((result) => {
+                if(result.data?.data?.deleteContactEventsBeforeDate) {
+                    deleteIrrelevantEventsLogger.info('Deleting contact events finished with success', Severity.LOW);
+                } else {
+                    deleteIrrelevantEventsLogger.error(`Deleting contact events finished with errors: ${result.data}`, Severity.LOW);
                     alertError(deletingContactEventsErrorMsg);
                     setDidDeletingContactEventsSucceed(false);
-                }).finally(() => {
-                    setIsLoading(false);
-                })
-            }
+                }
+            }).catch(err => {
+                deleteIrrelevantEventsLogger.error(`Failed to delete irrelevant contact events: ${err}`, Severity.LOW);
+                alertError(deletingContactEventsErrorMsg);
+                setDidDeletingContactEventsSucceed(false);
+            }).finally(() => {
+                setIsLoading(false);
+            })
         }
     }
 
-    const saveClinicalDetailsToDB = (clinicalDetails: ClinicalDetailsData, validationDate: Date, id: number) => {
+    const saveClinicalDetailsToDB = (clinicalDetails: ClinicalDetailsData, id: number) => {
         const saveClinicalDetailsLogger = logger.setup('Saving clinical details tab');
         saveClinicalDetailsLogger.info('launching the server request', Severity.LOW);
         setIsLoading(true);
@@ -250,17 +233,26 @@ const useClinicalDetails = (parameters: useClinicalDetailsIncome): useClinicalDe
             })
     }
 
-    const saveClinicalDetailsAndDeleteContactEvents = (clinicalDetails: ClinicalDetailsData, validationDate: Date, id: number): void => {
+    const alertSymptomsDatesChange = () =>
+        alertWarning('האם אתה בטוח שתרצה לשנות את שדה התסמינים? שינוי זה יגרום למחיקת האירועים והמגעים הקיימים בימים שימחקו', {
+            showCancelButton: true,
+            cancelButtonText: 'בטל',
+            cancelButtonColor: theme.palette.error.main,
+            confirmButtonColor: theme.palette.primary.main,
+            confirmButtonText: 'כן, המשך'
+        });
+
+    const saveClinicalDetailsAndDeleteContactEvents = (clinicalDetails: ClinicalDetailsData, id: number): void => {
         if(didSymptomsDateChangeOccur) {
             alertSymptomsDatesChange().then(result => {
                 if(result.isConfirmed) {
                     deleteIrrelevantContactEvents(clinicalDetails.symptomsStartDate, clinicalDetails.doesHaveSymptoms);
                     didDeletingContactEventsSucceed &&
-                        saveClinicalDetailsToDB(clinicalDetails, validationDate, id);
+                        saveClinicalDetailsToDB(clinicalDetails, id);
                 }
             })
         } else {
-            saveClinicalDetailsToDB(clinicalDetails, validationDate, id);
+            saveClinicalDetailsToDB(clinicalDetails, id);
         }
     }
